seeder added

+++ controller/root/admin.js
@@ -1,13 +1,17 @@
 const adminService = require("../../services/admin");
 const AdminRes = require("../../services/responses/AdminCreated");
 const dict = require("../../resources/dict");
+const Can = require("../../services/can/can");
 
 exports.store = async (req, res, next) => {
     try {
-        const storedAdmin = await adminService.insertAdmin(req);
-        let response = new AdminRes(200, "success", storedAdmin);
-        if (storedAdmin != "") res.status(200).send(response.handler());
-        else res.status(500).send("admin registration failed");
+        const storedAdminResponse = await adminService.insertAdmin(req);
+        if (storedAdminResponse != "") {
+            let response = new AdminRes(200, "success", storedAdminResponse);
+            return res.status(200).send(response.handler());
+        }
+        let response = new AdminRes(404, "fail", "failed");
+        return res.status(404).send(response.handler());
     } catch (e) {
         return res.status(500).send(e);
     }
@@ -15,6 +19,12 @@ exports.store = async (req, res, next) => {
 
 exports.update = async (req, res, next) => {
     try {
+        const can = await Can.can(req.admin.roleId, "update admin");
+        if (!can) {
+            let response = new AdminRes(403, "fail", "notAllowed");
+            return res.status(403).send(response.handler());
+        }
+
         const updatedAdminResponse = await adminService.updateAdmin(req);
         if (updatedAdminResponse === "adminNotFound") {
             let response = new AdminRes(404, "fail", "adminNotFound");
@@ -34,6 +44,12 @@ exports.update = async (req, res, next) => {
 };
 exports.destroy = async (req, res, next) => {
     try {
+        const can = await Can.can(req.admin.roleId, "delete admin");
+        if (!can) {
+            let response = new AdminRes(403, "fail", "notAllowed");
+            return res.status(403).send(response.handler());
+        }
+
         const destroyAdminResult = await adminService.destroyAdmin(req);
         if (destroyAdminResult === true) {
             let response = new AdminRes(200, "success", dict.successfulRemove);
@@ -50,6 +66,12 @@ exports.destroy = async (req, res, next) => {
 };
 exports.index = async (req, res, next) => {
     try {
+        const can = await Can.can(req.admin.roleId, "read admin");
+        if (!can) {
+            let response = new AdminRes(403, "fail", "notAllowed");
+            return res.status(403).send(response.handler());
+        }
+
         const adminIndexResponse = await adminService.indexAdmins();
         let response = new AdminRes(200, "success", adminIndexResponse);
